Move exception handler to worker
