+++ samples/sample-chat-app/src/main/AndroidManifest.xml
@@ -31,6 +31,21 @@
       </intent-filter>
     </service>
     <!-- [END firebase_service] -->
+
+    <!-- Disable the default initializer of  WorkManager so that we could override it in MyAppConfiguration  -->
+    <provider
+        android:name="androidx.startup.InitializationProvider"
+        android:authorities="${applicationId}.androidx-startup"
+        android:exported="false"
+        tools:node="merge">
+      <!-- If you are using androidx.startup to initialize other components -->
+      <meta-data
+          android:name="androidx.work.WorkManagerInitializer"
+          android:value="androidx.startup"
+          tools:node="remove" />
+    </provider>
+    <!--  End of Disabling  default initializer of  WorkManager -->
+
   </application>
 
 </manifest>
