+++ sdk/communication/azure-communication-chat/src/main/AndroidManifest.xml
@@ -11,21 +11,7 @@
         from the unit-test of other android packages; hence the override for minSdkVersion.
     -->
   <uses-sdk tools:overrideLibrary="com.azure.core.test" />
-  <application>
-    <!-- Disable the default initializer of  WorkManager so that we could override it in MyAppConfiguration  -->
-    <provider
-        android:name="androidx.startup.InitializationProvider"
-        android:authorities="${applicationId}.androidx-startup"
-        android:exported="false"
-        tools:node="merge">
-      <!-- If you are using androidx.startup to initialize other components -->
-      <meta-data
-          android:name="androidx.work.WorkManagerInitializer"
-          android:value="androidx.startup"
-          tools:node="remove" />
-    </provider>
-    <!--  End of Disabling  default initializer of  WorkManager -->
-  </application>
+
 
 
 </manifest>
