+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/ChatClient.java
@@ -223,11 +223,10 @@ public final class ChatClient {
      * If there's an error during registration initialization or renewal,
      * push notifications will be disabled and {@code errorHandler} will be called.
      * @param deviceRegistrationToken Device registration token obtained from the FCM SDK.
-     * @param errorHandler error handler callback for registration failures
      * @throws RuntimeException if push notifications failed to start.
      */
-    public void startPushNotifications(String deviceRegistrationToken, Consumer<Throwable> errorHandler) {
-        client.startPushNotifications(deviceRegistrationToken, errorHandler);
+    public void startPushNotifications(String deviceRegistrationToken) {
+        client.startPushNotifications(deviceRegistrationToken);
     }
 
     /**
