+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/implementation/notifications/fcm/RegistrationRenewalWorkerFactory.java
@@ -12,17 +12,22 @@ import androidx.work.WorkerParameters;
 
 import com.azure.android.communication.common.CommunicationTokenCredential;
 
+import java9.util.function.Consumer;
+
 public class RegistrationRenewalWorkerFactory extends WorkerFactory {
     private CommunicationTokenCredential communicationTokenCredential;
 
-    public RegistrationRenewalWorkerFactory(CommunicationTokenCredential communicationTokenCredential) {
+    private Consumer<Throwable> exceptionHandler;
+
+    public RegistrationRenewalWorkerFactory(CommunicationTokenCredential communicationTokenCredential, Consumer<Throwable> exceptionHandler) {
         this.communicationTokenCredential = communicationTokenCredential;
+        this.exceptionHandler = exceptionHandler;
     }
 
     @Nullable
     @Override
     public ListenableWorker createWorker(@NonNull Context appContext, @NonNull String workerClassName, @NonNull WorkerParameters workerParameters) {
 
-        return new RegistrationRenewalWorker(appContext, workerParameters, communicationTokenCredential);
+        return new RegistrationRenewalWorker(appContext, workerParameters, communicationTokenCredential, exceptionHandler);
     }
 }
